+++ src/main/java/com/jeyofdev/yellow_berry/domain/wishlist/WishlistService.java
@@ -3,6 +3,7 @@ package com.jeyofdev.yellow_berry.domain.wishlist;
 import com.jeyofdev.yellow_berry.core.classes.AbstractDomainService;
 import com.jeyofdev.yellow_berry.core.constant.ConfirmMessage;
 import com.jeyofdev.yellow_berry.core.constant.ErrorMessage;
+import com.jeyofdev.yellow_berry.domain.cart.Cart;
 import com.jeyofdev.yellow_berry.domain.product.Product;
 import com.jeyofdev.yellow_berry.domain.product.ProductRepository;
 import com.jeyofdev.yellow_berry.domain.profile.Profile;
@@ -10,9 +11,12 @@ import com.jeyofdev.yellow_berry.domain.profile.ProfileService;
 import com.jeyofdev.yellow_berry.exception.AlreadyTakenException;
 import com.jeyofdev.yellow_berry.exception.AlreadyAssociatedException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.text.MessageFormat;
+import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
@@ -39,41 +43,58 @@ public class WishlistService extends AbstractDomainService<WishList, WishlistRep
     }
 
     public WishList save(UUID profileId, WishList wishlist) {
+        String username  = SecurityContextHolder.getContext().getAuthentication().getName();
         Profile profile = profileService.findById(profileId);
 
-        if (profile.getWishlist() != null) {
-            throw new AlreadyAssociatedException(MessageFormat.format(ErrorMessage.ALREADY_ASSOCIATED, "profile", "wishlist"));
+        if (username.equals(profile.getUser().getUsername())) {
+            if (profile.getWishlist() != null) {
+                throw new AlreadyAssociatedException(MessageFormat.format(ErrorMessage.ALREADY_ASSOCIATED, "profile", "wishlist"));
+            }
+
+            if (wishlistRepository.existsByName(wishlist.getName())) {
+                throw new AlreadyTakenException(MessageFormat.format(ErrorMessage.ALREADY_TAKEN, entityName, "name", wishlist.getName()));
+            }
+
+            wishlist.setProfile(profile);
+            profile.setWishlist(wishlist);
+
+            return wishlistRepository.save(wishlist);
+        } else {
+            throw new AccessDeniedException(ErrorMessage.LIMIT_ACCESS);
         }
-
-        if (wishlistRepository.existsByName(wishlist.getName())) {
-            throw new AlreadyTakenException(MessageFormat.format(ErrorMessage.ALREADY_TAKEN, entityName, "name", wishlist.getName()));
-        }
-
-
-        wishlist.setProfile(profile);
-        profile.setWishlist(wishlist);
-
-        return wishlistRepository.save(wishlist);
     }
 
     public WishList updateById(UUID wishlistId, WishList updatedWishlist) {
+        String username  = SecurityContextHolder.getContext().getAuthentication().getName();
         WishList existingWishlist = findById(wishlistId);
-        existingWishlist.setName(updatedWishlist.getName() != null ? updatedWishlist.getName() : existingWishlist.getName());
 
-        return wishlistRepository.save(existingWishlist);
+        if (username.equals(existingWishlist.getProfile().getUser().getUsername())) {
+            existingWishlist.setName(updatedWishlist.getName() != null ? updatedWishlist.getName() : existingWishlist.getName());
+
+            return wishlistRepository.save(existingWishlist);
+        } else {
+            throw new AccessDeniedException(ErrorMessage.LIMIT_ACCESS);
+        }
     }
 
     public String deleteById(UUID wishlistId) {
+        String username  = SecurityContextHolder.getContext().getAuthentication().getName();
+        String roles  = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
         WishList wishlist = findById(wishlistId);
-        List<Product> productList = productRepository.findByWishlist(wishlist);
 
-        for (Product product : productList) {
-            product.getWishlists().remove(wishlist);
-            productRepository.save(product);
+        if (username.equals(wishlist.getProfile().getUser().getUsername()) || roles.equals("[ROLE_ADMIN]")) {
+            List<Product> productList = productRepository.findByWishlist(wishlist);
+
+            for (Product product : productList) {
+                product.getWishlists().remove(wishlist);
+                productRepository.save(product);
+            }
+
+            wishlistRepository.deleteById(wishlistId);
+
+            return ConfirmMessage.WISHLIST_DELETE;
+        } else {
+            throw new AccessDeniedException(ErrorMessage.LIMIT_ACCESS);
         }
-
-        wishlistRepository.deleteById(wishlistId);
-
-        return ConfirmMessage.WISHLIST_DELETE;
     }
 }
